% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognito_server.R
\name{cognito_server}
\alias{cognito_server}
\title{Cognito Shiny Module Server}
\usage{
cognito_server(
  input,
  output,
  session,
  with_cookie = FALSE,
  cookiename = "cognitor",
  cookie_expire = 7
)
}
\arguments{
\item{input}{- Shiny input}

\item{output}{- Shiny Output}

\item{session}{- Shiny Session}

\item{with_cookie}{- Create a own cookie when is authenticated in Cognito.}

\item{cookiename}{- name for cookie}

\item{cookie_expire}{- Expiration time for cookie}
}
\value{
reactiveValues (isLogged and userdata) and a callback function to do logout in Cognito.
}
\description{
A shiny server function to implement Cognito Authentication in your shiny app.
}
\author{
Pablo Pagnone
}
