% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognito_ui.R
\name{cognito_ui}
\alias{cognito_ui}
\title{Cognito UI
Offer a authentication with Amazon Cognito
This UI load the required JS.}
\usage{
cognito_ui(
  id,
  logout_btn = "Log out",
  logout_btn_class = "btn-danger btn-cognito-logout"
)
}
\arguments{
\item{id}{character - Namespace ID}

\item{logout_btn}{character - Text for Logout Button}

\item{logout_btn_class}{character - Class for logout button}
}
\description{
Cognito UI
Offer a authentication with Amazon Cognito
This UI load the required JS.
}
