% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie_server.R
\name{cookie_server}
\alias{cookie_server}
\title{Cookie Shiny Module Server}
\usage{
cookie_server(
  input,
  output,
  session,
  cookie_name = "user",
  cookie_expire = 0.5
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{cookie_name}{character - Name of cookie to create}

\item{cookie_expire}{numeric - Expiration timeof cookie}
}
\value{
list of reactive element and function to manage cookies
}
\description{
This server module returns a list with methods to interact with cookie created via JS.
You have:
}
\details{
\itemize{
\item getCookie - Reactive function, returns the content of cookie if exist. Else return FALSE.
\item setCookie - Sets the content for a cookie. (Required param: list())
\item rmCookie -  Removes a cookie
}
}
\examples{
cookie_server(NULL,NULL,NULL)
}
\author{
Pablo Pagnone
}
