% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie_ui.R
\name{cookie_ui}
\alias{cookie_ui}
\title{Cookie Shiny Module UI}
\usage{
cookie_ui(id)
}
\arguments{
\item{id}{character}
}
\value{
a Shiny UI
}
\description{
This UI module load the required js methods to create/remove/get cookie in browser.
}
\examples{
cookie_ui("cookie")
}
