% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Get Cognito config from .yml}
\usage{
get_config()
}
\value{
list|FALSE
}
\description{
Return all required configuration from file .yml to connect with Amazon Cognito instance.
}
\author{
Pablo Pagnone
}
