% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_user.R
\name{get_info_user}
\alias{get_info_user}
\title{Get Cognito Information of user via oauth}
\usage{
get_info_user(token, cognito_config)
}
\arguments{
\item{token}{string You can obtain with get_cognito_token_access()}

\item{cognito_config}{list List obtained with get_cognito_config()}
}
\value{
list|FALSE - If request is success return list with info about user. If configuration of token is invalid return FALSE.
}
\description{
Get Cognito Information of user via oauth
}
\examples{
get_info_user('your-token', get_config())

}
\author{
Pablo Pagnone
}
