% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token_access.R
\name{get_token_access}
\alias{get_token_access}
\title{Get Cognito Tokens via oauth}
\usage{
get_token_access(code, cognito_config)
}
\arguments{
\item{code}{character - Code Access received from Amazon Cognito}

\item{cognito_config}{list - List obtained with get_cognito_config()}
}
\value{
list|FALSE
}
\description{
With code received via url, get tokens for user via oauth.
}
\author{
Pablo Pagnone
}
