% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_auth_redirect.R
\name{get_url_auth_redirect}
\alias{get_url_auth_redirect}
\title{Get Cognito URL for authentication redirect}
\usage{
get_url_auth_redirect(cognito_config, session = getDefaultReactiveDomain())
}
\arguments{
\item{cognito_config}{list Obtained with get_cognito_config()}

\item{session}{Shiny session Will be used to keep the params in urls when redirection is done.}
}
\value{
character|FALSE
}
\description{
Return Url where user is redirect if is not logged yet.
}
\examples{
get_url_auth_redirect(get_config())
}
\author{
Pablo Pagnone
}
