% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_logout_redirect.R
\name{get_url_logout_redirect}
\alias{get_url_logout_redirect}
\title{Get Cognito URL for Logout redirect.}
\usage{
get_url_logout_redirect(cognito_config)
}
\arguments{
\item{cognito_config}{list Obtained with get_cognito_config()}
}
\value{
character|FALSE
}
\description{
Return Url where user is redirect when is logout.
}
\author{
Pablo Pagnone
}
