% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_server.R
\name{logout_server}
\alias{logout_server}
\title{Logout Server
Receive a reactive element (isLogged), when is TRUE the logout button and information about is shown.}
\usage{
logout_server(
  input,
  output,
  session,
  isLogged = reactive(FALSE),
  textlogged = "You are logged in"
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{isLogged}{reactive}

\item{textlogged}{character}
}
\value{
reactive (the logout button)
}
\description{
Logout Server
Receive a reactive element (isLogged), when is TRUE the logout button and information about is shown.
}
