% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognito_server.R
\name{cognito_server}
\alias{cognito_server}
\title{Cognito Server}
\usage{
cognito_server(input, output, session)
}
\arguments{
\item{input}{- Shiny input}

\item{output}{- Shiny Output}

\item{session}{- Shiny Session}
}
\value{
reactiveValues (isLogged and userdata) and callback function to do logout in Cognito.
}
\description{
Return shiny server for Cognito Module.
}
