% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognito_ui.R
\name{cognito_ui}
\alias{cognito_ui}
\title{Cognito Shiny Module UI}
\usage{
cognito_ui(id)
}
\arguments{
\item{id}{character - Namespace ID}
}
\value{
A Shiny UI
}
\description{
A Shiny UI function to be used to implement Amazon Cognito in your shiny app.
This UI loads the required JS.
}
\author{
Pablo Pagnone
}
