% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie_server.R
\name{cookie_server}
\alias{cookie_server}
\title{Cookie Shiny Module Server}
\usage{
cookie_server(
  input,
  output,
  session,
  cookie_name = "user",
  cookie_expire = 0.5
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{cookie_name}{character - Name of cookie to create}

\item{cookie_expire}{numeric - Expiration timeof cookie}
}
\value{
list with reactive element and function to manage cookie
}
\description{
This server module return a list with methods to interact with cookie created via JS.
You have:
getCookie - Reactive function, return the content of cookie if exist. Else return FALSE.
setCookie - Allow to set the content for a cookie. (Required param: list())
rmCookie - Allow to remove a cookie.
}
\author{
Pablo Pagnone
}
