% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_ui.R
\name{logout_ui}
\alias{logout_ui}
\title{Logout UI
Return a UI with a logout button and information about user logged.
By default is hidden and is show with reactive element from logoutServer}
\usage{
logout_ui(id, textbutton = "Log out", classbutton = "btn-logout btn-danger")
}
\arguments{
\item{id}{character}

\item{textbutton}{character}

\item{classbutton}{character}
}
\value{
Shiny UI
}
\description{
Logout UI
Return a UI with a logout button and information about user logged.
By default is hidden and is show with reactive element from logoutServer
}
